plugins: move move-drag-interface into a shared library
