+++ plugins/tile/meson.build
@@ -3,6 +3,7 @@ tile = shared_module('simple-tile',
         include_directories: [wayfire_api_inc, wayfire_conf_inc, plugins_common_inc, grid_inc, wobbly_inc, ipc_include_dirs],
         dependencies: [wlroots, pixman, wfconfig, json, plugin_pch_dep],
         cpp_args: plugin_pch_args,
+        link_with: [move_drag_interface],
         install: true,
         install_dir: join_paths(get_option('libdir'), 'wayfire'))
